requirements.txt
fastapi~=0.115.5
uvicorn
bcrypt
sqlmodel~=0.0.22
SQLAlchemy~=2.0.36
python-dotenv
pyjwt
passlib